goby_backend/requirements.txt
Django==4.1.1
djangorestframework==3.13.1
mysqlclient==2.1.1
django-mysql==4.7.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-allauth==0.51.0
dj-rest-auth==2.2.5
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4